put current command into subfolder
